+++ nakone-law/src/App.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 import "./App.css";
 import Header from './components/Header/Header';
 import Navigation from './components/Navigation/Navigation';
-import Banner from './Banner/Banner';
-import Text from './Text/Text';
-import Buttom from './Buttom/Buttom';
+import Banner from './components/Banner/Banner';
+import Button from './components/Button/Button';
+// import Carousel from './components/Carousel/Carousel';
 
 const App = () => (
   <div>
     <Header />
     <Navigation />
     <Banner />
-    <Text />
-    <Buttom />
+    <Button />
+    {/* <Carousel /> */}
   </div>
 );
