--- nakone-law/src/Banner/Banner.css
@@ -1,7 +1,0 @@
-.banner{
-  display: flex;
-  /* width: 50%; */
-  justify-content: center;
-  align-items: center;
-  margin: 0 auto;
-}
