--- nakone-law/src/Buttom/Buttom.jsx
@@ -1,9 +1,0 @@
-import './Buttom.css';
-
-const Buttom = () => {
-  <button className="buttom">
-    CONSULTATION
-  </button>;
-}
-
-export default Buttom;
