--- nakone-law/src/Text/Text.css
@@ -1,12 +1,0 @@
-.text {
-  display: flex;
-  color: #213555;
-  text-align: center;
-  align-self: stretch;
-/* Main text (book) */
-  font-family: PP Neue Montreal;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 150%; /* 24px */
-}
