--- nakone-law/src/Text/Text.jsx
@@ -1,9 +1,0 @@
-import './Text.css';
-
-const Text = () => {
-  <h3 className="text">
-    Legal services in the field of immigration law in the United States
-  </h3>;
-}
-
-export default Text;
