+++ nakone-law/src/components/Carousel/Carousel.css
@@ -1,0 +1,5 @@
+.carousel {
+  width: 297.477px;
+  height: 400px;
+  flex-shrink: 0;
+}
